i59: Extract Bottom Sheet to component
