+++ src/library/components/index.js
@@ -1,5 +1,6 @@
 export { Menu, MenuDivider, MenuItem } from './Menu';
 export { default as ActivityOverlay } from './ActivityOverlay';
+export { default as BottomSheet } from './BottomSheet';
 export { default as FAB } from './FAB';
 export { default as HomeBar } from './HomeBar';
 export { default as IconButton } from './IconButton';
@@ -9,4 +10,3 @@ export { default as OnboardingButtons } from './OnboardingButtons';
 export { default as Scrim } from './Scrim';
 export { default as TwitterAuth } from './TwitterAuth';
 export { default as WorldMap } from './WorldMap';
-
